Fix : ScriptableObjectMapReader and SerializableMapReader methods
